+++ variables.tf
@@ -11,18 +11,19 @@ variable "services" {
   )
 }
 
-# Common variables
+# PacketFabric Cloud Router
 variable "name" {
   description = "The base name all Network services created in PacketFabric, Google and AWS."
   type        = string
-}
-variable "labels" {
-  description = "The labels to be assigned to the PacketFabric Cloud Router and Cloud Router Connections."
-  type        = list(string)
-  default     = ["terraform-cts"]
+  default     = null
 }
 
-# PacketFabric Cloud Router
+variable "cr_id" {
+  description = "The Cloud Router Circuit ID."
+  type        = string
+  default     = null
+}
+
 variable "asn" {
   description = "The Autonomous System Number (ASN) for the PacketFabric Cloud Router."
   type        = number
@@ -32,7 +33,7 @@ variable "asn" {
 variable "capacity" {
   description = "The capacity of the PacketFabric Cloud Router."
   type        = string
-  default     = "10Gbps"
+  default     = ">100Gbps"
 }
 
 variable "regions" {
@@ -41,15 +42,18 @@ variable "regions" {
   default     = ["US"]
 }
 
-variable "cr_id" {
-  default     = null
-  description = "The Cloud Router Circuit ID created by this module automatically."
+variable "labels" {
+  description = "The labels to be assigned to the PacketFabric Cloud Router and Cloud Router Connections."
+  type        = list(string)
+  default     = ["terraform-cts"]
 }
 
 # PacketFabric Cloud Router Connection AWS
 variable "aws_cloud_router_connections" {
-  description = "An object representing the AWS Cloud Router Connections."
-  type = object({
+  description = "List of objects representing the AWS Cloud Router Connections."
+  type = list(object({
+    name       = optional(string)
+    labels     = optional(list(string))
     aws_region = string
     aws_vpc_id = string
     aws_asn1   = optional(number)
@@ -62,14 +66,21 @@ variable "aws_cloud_router_connections" {
       type   = string
     })))
     bgp_prefixes_match_type = optional(string)
-  })
+  }))
   default = null
+
+  validation {
+    condition     = length(coalesce(var.aws_cloud_router_connections, [])) <= 3
+    error_message = "You must provide no more than three AWS Cloud Router Connections."
+  }
 }
 
 # PacketFabric Cloud Router Conection Google
 variable "google_cloud_router_connections" {
-  description = "A object representing the Google Cloud Router Connections."
-  type = object({
+  description = "List of objects representing the Google Cloud Router Connections."
+  type = list(object({
+    name           = optional(string)
+    labels         = optional(list(string))
     google_project = string
     google_region  = string
     google_network = string
@@ -82,27 +93,38 @@ variable "google_cloud_router_connections" {
       type   = string
     })))
     bgp_prefixes_match_type = optional(string)
-  })
+  }))
   default = null
 }
 
-# # PacketFabric Cloud Router Conection Azure -- not yet available open an issue on github
-# variable "azure_cloud_router_connections" {
-#   description = "An object representing the Azure Cloud Router Connections."
-#   type = object({
-#     azure_region  = string
-#     azure_vnet_id = string
-#     azure_pop     = string
-#     azure_speed   = optional(string)
-#     redundant     = optional(bool)
-#     bgp_prefixes = optional(list(object({
-#       prefix = string
-#       type   = string
-#     })))
-#     bgp_prefixes_match_type = optional(string)
-#   })
-#   default = null
-# }
+# PacketFabric Cloud Router Conection Azure -- not yet available open an issue on github
+variable "azure_cloud_router_connections" {
+  description = "List of objects representing the Azure Cloud Router Connections."
+  type = list(object({
+    name                  = optional(string)
+    labels                = optional(list(string))
+    azure_region          = string
+    azure_resource_group  = string
+    azure_vnet            = string
+    azure_pop             = string
+    azure_speed           = optional(string)
+    azure_subscription_id = optional(string)
+    skip_gateway          = optional(bool)
+    redundant             = optional(bool)
+    bgp_prefixes = optional(list(object({
+      prefix = string
+      type   = string
+    })))
+    bgp_prefixes_match_type = optional(string)
+    provider                = optional(string) # Use "Packet Fabric Test" for internal PF dev testing
+  }))
+  default = null
+
+  validation {
+    condition     = length(coalesce(var.azure_cloud_router_connections, [])) <= 1
+    error_message = "You must provide no more than one Azure Cloud Router Connection."
+  }
+}
 
 variable "aws_in_prefixes" {
   description = "The Allowed Prefixes from AWS, by default will get all AWS VPC subnets. You can also add additional ones."
@@ -112,7 +134,7 @@ variable "google_in_prefixes" {
   description = "The Allowed Prefixes from Google Cloud, by default will get all Google VPC subnets. You can also add additional ones."
   default     = []
 }
-# variable "azure_in_prefixes" {
-#   description = "The Allowed Prefixes from Azure Cloud, by default will get all Azure VNet subnets. You can also add additional ones."
-#   default     = []
-# }
+variable "azure_in_prefixes" {
+  description = "The Allowed Prefixes from Azure Cloud, by default will get all Azure VNet subnets. You can also add additional ones."
+  default     = []
+}