+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 0.3.0  (May 25, 2023)
+
+FEATURES:
+
+* Update Cloud Router Terraform Module to version 0.3.0 (#2)
+    * Adding Azure standalone or redundant Cloud Router Connection support
+    * Adding name and labels attributes to individual Cloud Router Conections
+    * Adding support of multiple Google and AWS Cloud Router Connections
+    * Adding option to specify an existing PacketFabric Cloud Router
+
 ## 0.1.1  (May 8, 2023)
 
 IMPROVEMENTS/ENHANCEMENTS: